Battery: move detection code into `src/detection`
Prepare for macOS implementation
