+++ src/detection/battery/battery_apple.c
@@ -1,0 +1,7 @@
+#include "fastfetch.h"
+#include "battery.h"
+
+const char* ffDetectBatteryImpl(FFinstance* instance, FFlist* results) {
+    FF_UNUSED(instance, results);
+    return "Unimplemented";
+}
